refactor(ir.core.module): say it once, where it belongs
Summary:
- Fold the shared-child alias invariant into the module-surface tests it belongs
  with and drop the file that carried it alone; what remains of its reasoning is
  the one line about why `copy.copy` builds an alias construction prevents.
- Trim the docstrings and comments added with `load`/`LoadedModule` to the
  behaviour, leaving the contract to docs/spec/runtime.md §1.1.2.
